AbstractAxis

Currently this functionality is exerimental. If the AbstractAxis interface proves useful it will likely be moved to a different package.

Indexing

The standard syntax for indexing doesn't change at all.

julia> using StaticRanges

julia> sa = SimpleAxis(1:10)
SimpleAxis(1:10)

julia> sa[2]
2

julia> sa[>(2)]
SimpleAxis(3:10)

julia> a = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

julia> a[2]
2

julia> a[2:3]
Axis(2:3 => 2:3)

But now we can also use functions to index by the the keys of an AbstractAxis.

julia> using StaticRanges

julia> a = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> a[1]
1

julia> a[isequal(2.0)]
1

julia> a[>(2)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[>(2.0)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[and(>(2.0), <(8.0))]
Axis(3.0:1.0:7.0 => 2:6)

Benchmarks

Indexing CartesianAxes is comparable to that of CartesianIndices

julia> using StaticRanges, BenchmarkTools

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> @btime getindex(cartaxes, 2, 2)
  20.848 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartinds, 2, 2)
  22.317 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartaxes, ==(3.0), 2)
  444.374 ns (7 allocations: 416 bytes)
CartesianIndex(2, 2)

Indexing LinearAxes is comparable to that of LinearIndices

julia> using StaticRanges, BenchmarkTools

julia> linaxes = LinearAxes((Axis(1.0:4.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> @btime getindex(linaxes, 2, 2)
  18.275 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(lininds, 2, 2)
  18.849 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(linaxes, ==(3.0), 2)
  381.098 ns (6 allocations: 384 bytes)
7

Chaining filters

StaticRanges.and — Function

and(x, y)

Synonymous with bitwise & operator but may be used to chain multiple Fix1 or Fix2 operations.

Examples

julia> using StaticRanges

julia> and(<(5), true)(1)
true

julia> and(<(5), false)(1)
false

julia> and(and(<(5), >(1)), >(2))(3)
true

StaticRanges.or — Function

or(x, y)

Synonymous with bitwise | operator but may be used to chain multiple Fix1 or Fix2 operations.

Examples

julia> using StaticRanges

julia> or(<(5), true)(1)
true

julia> or(<(5), false)(1)
true

julia> or(or(<(5), >(1)), >(2))(3)
true

Reindexing

StaticRanges.reindex — Function

reindex(a::AbstractAxis, inds::AbstractVector{Integer}) -> AbstractAxis

Returns and index of the same type as a where the keys the new keys are constructed by indexing into the keys of a with inds and the values have the same starting value but a length matching inds.

Examples

julia> using StaticRanges

julia> x, y, z = Axis(1:10, 2:11), Axis(1:10), SimpleAxis(1:10);

julia>  reindex(x, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => 2:6)

julia> reindex(y, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => Base.OneTo(5))

julia> reindex(z, collect(1:2:10))
SimpleAxis(1:5)

StaticRanges.unsafe_reindex — Function

unsafe_reindex(a::AbstractAxis, inds::AbstractVector) -> AbstractAxis

Similar to reindex this function returns an index of the same type as a but doesn't check that inds is inbounds. New subtypes of AbstractAxis must implement a unique unsafe_reindex method.

See also: reindex

Examples

julia> using StaticRanges

julia> StaticRanges.unsafe_reindex(SimpleAxis(OneToMRange(10)), 1:5)
SimpleAxis(OneToMRange(5))

julia> StaticRanges.unsafe_reindex(SimpleAxis(OneToSRange(10)), 1:5)
SimpleAxis(OneToSRange(5))

Types

StaticRanges.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

StaticRanges.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using StaticRanges

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

StaticRanges.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using StaticRanges

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
3

julia> x[==(2)]
2

julia> x[2] == x[==(3)]
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

Interface

StaticRanges.values_type — Function

values_type(::AbstractAxis)

StaticRanges.keys_type — Function

keys_type(::AbstractAxis)

Swapping Axes

StaticRanges.drop_axes — Function

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using StaticRanges

julia> axs = (Axis(1:5), Axis(1:10));

julia> drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

StaticRanges.permute_axes — Function

permute_axes(x::AbstractArray, p::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, p::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using StaticRanges

julia> permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using StaticRanges

julia> permute_axes(rand(4))
(Base.OneTo(1), Base.OneTo(4))

julia> permute_axes((Axis(1:4),))
(Axis(1:1 => Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using StaticRanges

julia> permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

StaticRanges.reduce_axes — Function

reduce_axes(a, dims)

Returns the appropriate axes for a measure that reduces dimensions along the dimensions dims.

Example

julia> using StaticRanges

julia> reduce_axes(rand(2, 4), 2)
(Base.OneTo(2), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), (1,2))
(Base.OneTo(1), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), :)
()

julia> reduce_axes((Axis(1:4), Axis(1:4)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:1 => Base.OneTo(1)))

StaticRanges.reduce_axis — Function

reduce_axis(a)

Reduces axis a to single value. Allows custom index types to have custom behavior throughout reduction methods (e.g., sum, prod, etc.)

See also: reduce_axes

Example

julia> using StaticRanges

julia> reduce_axis(Axis(1:4))
Axis(1:1 => Base.OneTo(1))

julia> reduce_axis(1:4)
1:1

Matrix Multiplication and Axes

StaticRanges.matmul_axes — Function

matmul_axes(a, b) -> Tuple

Returns the appropriate axes for the return of a * b where a and b are a vector or matrix.

Examples

julia> using StaticRanges

julia> axs2, axs1 = (Axis(1:2), Axis(1:4)), (Axis(1:6),);

julia> matmul_axes(axs2, axs2)
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs1, axs2)
(Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs2, axs1)
(Axis(1:2 => Base.OneTo(2)),)

julia> matmul_axes(axs1, axs1)
()

julia> matmul_axes(rand(2, 4), rand(4, 2))
(Base.OneTo(2), Base.OneTo(2))

StaticRanges.inverse_axes — Function

inverse_axes(a::AbstractMatrix) = inverse_axes(axes(a))
inverse_axes(a::Tuple{I1,I2}) -> Tuple{I2,I1}

Returns the inverted axes of a, corresponding to the inv method from the LinearAlgebra package in the standard library.

Examples

julia> using StaticRanges

julia> inverse_axes(rand(2,4))
(Base.OneTo(4), Base.OneTo(2))

julia> inverse_axes((Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:2 => Base.OneTo(2)))

StaticRanges.covcor_axes — Function

covcor_axes(x, dim) -> NTuple{2}

Returns appropriate axes for a cov or var method on array x.

Examples

julia> using StaticRanges

julia> covcor_axes(rand(2,4), 1)
(Base.OneTo(4), Base.OneTo(4))

julia> covcor_axes((Axis(1:4), Axis(1:6)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> covcor_axes((Axis(1:4), Axis(1:4)), 1)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

Appending Axes

StaticRanges.append_axes — Function

append_axes(x, y)

Returns the axes for append(x, y).

julia> using StaticRanges

julia> append_axes(1:10, 1:10)
(Base.OneTo(20),)

StaticRanges.append_axes! — Function

append_axes!(x, y)

Returns the axes for append!(x, y).

StaticRanges.append_keys — Function

append_keys(x, y)

Returns the appropriate keys of and index within the operation append_axis(x, y)

See also: append_axis

StaticRanges.append_values — Function

append_values(x, y)

Returns the appropriate values of and index within the operation append_axis(x, y)

See also: append_axis

StaticRanges.append_axis — Function

append_axis(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis method.

Examples

julia> using StaticRanges

julia> append_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> append_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

StaticRanges.append_axis! — Function

append_axis!(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis! method.

Examples

julia> using StaticRanges

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> append_axis!(x, y);

julia> length(x)
20

julia> append_axis!(y, x);

julia> length(y)
30

Concatenating Axes

StaticRanges.hcat_axes — Function

hcat_axes(x, y) -> Tuple

Returns the appropriate axes for hcat(x, y).

Examples

julia> using StaticRanges

julia> hcat_axes((Axis(1:4), Axis(1:2)), (Axis(1:4), Axis(1:2)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1; 2; 3; 4; 5], [6 7; 8 9; 10 11; 12 13; 14 15];

julia> hcat_axes(a, b) == axes(hcat(a, b))
true

StaticRanges.vcat_axes — Function

vcat_axes(x, y) -> Tuple

Returns the appropriate axes for vcat(x, y).

Examples

julia> using StaticRanges

julia> vcat_axes((Axis(1:2), Axis(1:4)), (Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1 2 3 4 5], [6 7 8 9 10; 11 12 13 14 15];

julia> vcat_axes(a, b) == axes(vcat(a, b))
true

StaticRanges.cat_axis — Function

cat_axis(x, y)

Returns the concatenation of the axes x and y. New subtypes of AbstractAxis must implement a unique cat_axis method.

StaticRanges.cat_values — Function

cat_values(x, y)

Returns the appropriate values of the x and y index within the operation cat_axis(x, y)

See also: cat_axis

StaticRanges.cat_keys — Function

cat_keys(x, y)

Returns the appropriate keys of the x and y index within the operation cat_axis(x, y)

See also: cat_axis

Resizing Axes

StaticRanges.resize_first — Function

resize_first(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the first index to be of size n.

Examples

julia> using StaticRanges

julia> x = collect(1:5);

julia> resize_first(x, 2)
2-element Array{Int64,1}:
 4
 5

julia> resize_first(x, 7)
7-element Array{Int64,1}:
 -1
  0
  1
  2
  3
  4
  5

julia> resize_first(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

StaticRanges.resize_first! — Function

resize_first!(x, n::Integer)

Returns the collection x after growing or shrinking the first index to be of size n.

Examples

julia> using StaticRanges

julia> x = collect(1:5);

julia> resize_first!(x, 2);

julia> x
2-element Array{Int64,1}:
 4
 5

julia> resize_first!(x, 6);

julia> x
6-element Array{Int64,1}:
 0
 1
 2
 3
 4
 5

StaticRanges.resize_last — Function

resize_last(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the last index to be of size n.

Examples

julia> using StaticRanges

julia> x = collect(1:5);

julia> resize_last(x, 2)
2-element Array{Int64,1}:
 1
 2

julia> resize_last(x, 7)
7-element Array{Int64,1}:
 1
 2
 3
 4
 5
 6
 7

julia>  resize_last(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

StaticRanges.resize_last! — Function

resize_last!(x, n::Integer)

Returns the collection x after growing or shrinking the last index to be of size n.

Examples

julia> using StaticRanges

julia> x = collect(1:5);

julia> resize_last!(x, 2);

julia> x
2-element Array{Int64,1}:
 1
 2

julia> resize_last!(x, 5);

julia> x
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

StaticRanges.grow_first — Function

grow_first(x, n)

Returns a collection similar to x that grows by n elements from the first index.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first(mr, 2)
UnitMRange(-1:10)

StaticRanges.grow_first! — Function

grow_first!(x, n)

Returns the collection x after growing from the first index by n elements.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first!(mr, 2);

julia> mr
UnitMRange(-1:10)

StaticRanges.grow_last — Function

grow_last(x, n)

Returns a collection similar to x that grows by n elements from the last index.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last(mr, 2)
UnitMRange(1:12)

StaticRanges.grow_last! — Function

grow_last!(x, n)

Returns the collection x after growing from the last index by n elements.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last!(mr, 2);

julia> mr
UnitMRange(1:12)

StaticRanges.shrink_first — Function

shrink_first(x, n)

Returns a collection similar to x that shrinks by n elements from the first index.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_first(mr, 2)
UnitMRange(3:10)

StaticRanges.shrink_first! — Function

shrink_first!(x, n)

Returns the collection x after shrinking from the first index by n elements.

StaticRanges.shrink_last — Function

shrink_last(x, n)

Returns a collection similar to x that shrinks by n elements from the last index.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_first!(mr, 2);

julia> mr
UnitMRange(3:10)

StaticRanges.shrink_last! — Function

shrink_last!(x, n)

Returns the collection x after shrinking from the last index by n elements.

StaticRanges.next_type — Function

next_type(x::T)

Returns the immediately greater value of type T.

StaticRanges.prev_type — Function

prev_type(x::T)

Returns the immediately lesser value of type T.

Combine Indices

StaticRanges.combine_indices — Function

combine_indices(x, y)

Returns the combined axes of x and y for broadcasting operations.

StaticRanges.combine_index — Function

combine_index(x, y)

Returns the combination of x and y, creating a new index. New subtypes of AbstractAxis should implement a combine_index method.

StaticRanges.combine_values — Function

combine_values(x, y)

Returns the combination of the values of x and y, creating a new index. New subtypes of AbstractAxis may implement a unique combine_values method if needed. Default behavior is to use the return of promote_rule(x, y) for the type of the combined values.

StaticRanges.combine_keys — Function

combine_keys(x::AbstractAxis, y::AbstractAxis)

Returns the combination of the keys of x and y, creating a new index. New subtypes of AbstractAxis may implement a unique combine_keys method if needed. Default behavior is to use the return of promote_rule(x, y) for the type of the combined keys.