Appending Axes

StaticRanges.append_axes — Function

append_axes(x, y)

Returns the axes for append(x, y).

julia> using StaticRanges

julia> append_axes(1:10, 1:10)
(Base.OneTo(20),)

StaticRanges.append_axes! — Function

append_axes!(x, y)

Returns the axes for append!(x, y).

StaticRanges.append_keys — Function

append_keys(x, y)

Returns the appropriate keys of and index within the operation append_axis(x, y)

See also: append_axis

StaticRanges.append_values — Function

append_values(x, y)

Returns the appropriate values of and index within the operation append_axis(x, y)

See also: append_axis

StaticRanges.append_axis — Function

append_axis(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis method.

Examples

julia> using StaticRanges

julia> append_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> append_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

StaticRanges.append_axis! — Function

append_axis!(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis! method.

Examples

julia> using StaticRanges

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> append_axis!(x, y);

julia> length(x)
20

julia> append_axis!(y, x);

julia> length(y)
30