Axes to Arrays

Most of the methods up to this point are potentially useful for manipulating an axis independent of a parent structure that it may belong to. However, we want it to be easy to use an AbstractAxis in a variety of settings. The following methods are likely to only be useful when extending the use of a multidimensional arrays to use the AbstractAxis interface.

Swapping Axes

StaticRanges.drop_axes — Function

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using StaticRanges

julia> axs = (Axis(1:5), Axis(1:10));

julia> drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

StaticRanges.permute_axes — Function

permute_axes(x::AbstractArray, p::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, p::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using StaticRanges

julia> permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using StaticRanges

julia> permute_axes(rand(4))
(Base.OneTo(1), Base.OneTo(4))

julia> permute_axes((Axis(1:4),))
(Axis(1:1 => Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using StaticRanges

julia> permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

StaticRanges.reduce_axes — Function

reduce_axes(a, dims)

Returns the appropriate axes for a measure that reduces dimensions along the dimensions dims.

Example

julia> using StaticRanges

julia> reduce_axes(rand(2, 4), 2)
(Base.OneTo(2), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), (1,2))
(Base.OneTo(1), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), :)
()

julia> reduce_axes((Axis(1:4), Axis(1:4)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:1 => Base.OneTo(1)))

StaticRanges.reduce_axis — Function

reduce_axis(a)

Reduces axis a to single value. Allows custom index types to have custom behavior throughout reduction methods (e.g., sum, prod, etc.)

See also: reduce_axes

Example

julia> using StaticRanges

julia> reduce_axis(Axis(1:4))
Axis(1:1 => Base.OneTo(1))

julia> reduce_axis(1:4)
1:1

Matrix Multiplication and Statistics

StaticRanges.matmul_axes — Function

matmul_axes(a, b) -> Tuple

Returns the appropriate axes for the return of a * b where a and b are a vector or matrix.

Examples

julia> using StaticRanges

julia> axs2, axs1 = (Axis(1:2), Axis(1:4)), (Axis(1:6),);

julia> matmul_axes(axs2, axs2)
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs1, axs2)
(Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs2, axs1)
(Axis(1:2 => Base.OneTo(2)),)

julia> matmul_axes(axs1, axs1)
()

julia> matmul_axes(rand(2, 4), rand(4, 2))
(Base.OneTo(2), Base.OneTo(2))

julia> matmul_axes(CartesianAxes((2,4)), CartesianAxes((4, 2))) == matmul_axes(rand(2, 4), rand(4, 2))
true

StaticRanges.inverse_axes — Function

inverse_axes(a::AbstractMatrix) = inverse_axes(axes(a))
inverse_axes(a::Tuple{I1,I2}) -> Tuple{I2,I1}

Returns the inverted axes of a, corresponding to the inv method from the LinearAlgebra package in the standard library.

Examples

julia> using StaticRanges

julia> inverse_axes(rand(2,4))
(Base.OneTo(4), Base.OneTo(2))

julia> inverse_axes((Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:2 => Base.OneTo(2)))

StaticRanges.covcor_axes — Function

covcor_axes(x, dim) -> NTuple{2}

Returns appropriate axes for a cov or var method on array x.

Examples

julia> using StaticRanges

julia> covcor_axes(rand(2,4), 1)
(Base.OneTo(4), Base.OneTo(4))

julia> covcor_axes((Axis(1:4), Axis(1:6)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> covcor_axes((Axis(1:4), Axis(1:4)), 1)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))