Indexing

The standard syntax for indexing doesn't change at all.

julia> using StaticRanges

julia> sa = SimpleAxis(1:10)
SimpleAxis(1:10)

julia> sa[2]
2

julia> sa[>(2)]
SimpleAxis(3:10)

julia> a = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

julia> a[2]
2

julia> a[2:3]
Axis(2:3 => 2:3)

But now we can also use functions to index by the the keys of an AbstractAxis.

julia> using StaticRanges

julia> a = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> a[1]
1

julia> a[isequal(2.0)]
1

julia> a[>(2)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[>(2.0)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[>(2.0) & <(8.0)]
Axis(3.0:1.0:7.0 => 2:6)

Chaining filters

StaticRanges.and — Function

and(x, y)

Synonymous with bitwise & operator but may be used to chain multiple Fix1 or Fix2 operations.

StaticRanges.or — Function

or(x, y)

Synonymous with bitwise | operator but may be used to chain multiple Fix1 or Fix2 operations.

Reindexing

StaticRanges.reindex — Function

reindex(a::AbstractAxis, inds::AbstractVector{Integer}) -> AbstractAxis

Returns and index of the same type as a where the keys the new keys are constructed by indexing into the keys of a with inds and the values have the same starting value but a length matching inds.

Examples

julia> using StaticRanges

julia> x, y, z = Axis(1:10, 2:11), Axis(1:10), SimpleAxis(1:10);

julia>  reindex(x, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => 2:6)

julia> reindex(y, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => Base.OneTo(5))

julia> reindex(z, collect(1:2:10))
SimpleAxis(1:5)

StaticRanges.unsafe_reindex — Function

unsafe_reindex(a::AbstractAxis, inds::AbstractVector) -> AbstractAxis

Similar to reindex this function returns an index of the same type as a but doesn't check that inds is inbounds. New subtypes of AbstractAxis must implement a unique unsafe_reindex method.

See also: reindex

Examples

julia> using StaticRanges

julia> StaticRanges.unsafe_reindex(SimpleAxis(OneToMRange(10)), 1:5)
SimpleAxis(OneToMRange(5))

julia> StaticRanges.unsafe_reindex(SimpleAxis(OneToSRange(10)), 1:5)
SimpleAxis(OneToSRange(5))