Order Functions

Optimized for Order

Traits are used to conveniently characterize the order of ranges.

StaticRanges.is_forward — Function

is_forward(x) -> Bool

Returns true if x is sorted forward.

Examples

julia> using StaticRanges

julia> fr = 1:2:10
1:2:9

julia> rr = 10:-2:1
10:-2:2

julia> is_forward(fr)
true

julia> is_forward(rr)
false

StaticRanges.is_reverse — Function

is_reverse(x) -> Bool

Returns true if x is sorted in reverse.

Examples

julia> using StaticRanges

julia> fr = 1:2:10
1:2:9

julia> rr = 10:-2:1
10:-2:2

julia> is_reverse(fr)
false

julia> is_reverse(rr)
true

StaticRanges.order — Function

order(x) -> Ordering

Returns the ordering of x.

StaticRanges.is_ordered — Function

is_ordered(x) -> Bool

Returns true if x is ordered. is_ordered should return the same value that issorted would on x except it doesn't specify how it's sorted (e.g., forward, reverse, etc).

StaticRanges.ordmax — Function

ordmax(x) = ordmax(x, order(x))
ordmax(x::T, ::Ordering) -> T

Finds the maximum of x using information about its ordering.

StaticRanges.ordmin — Function

ordmin(x) = ordmin(x, order(x))
ordmin(x::T, ::Ordering) -> T

Finds the minimum of x using information about its ordering.

StaticRanges.find_max — Function

find_max(x)

Returns the index of the maximum value for x. Differes from findmax by accounting for any sorting.

StaticRanges.find_min — Function

find_min(x)

Returns the index of the minimum value for x. Differes from findmin by accounting for any sorting.

StaticRanges.is_within — Function

is_within(x, y) -> Bool

Returns true if all of x is found within y.

StaticRanges.gtmax — Function

gtmax(x, y) -> Bool

Returns true if the maximum of x is greater than that of y.

StaticRanges.ltmax — Function

ltmax(x, y) -> Bool

Returns true if the maximum of x is less than that of y.

StaticRanges.eqmax — Function

eqmax(x, y) -> Bool

Returns true if the maximum of x and y are equal.

StaticRanges.gtmin — Function

gtmin(x, y) -> Bool

Returns true if the minimum of x is greater than that of y.

StaticRanges.ltmin — Function

ltmin(x, y) -> Bool

Returns true if the minimum of x is less than that of y.

StaticRanges.eqmin — Function

eqmin(x, y) -> Bool

Returns true if the minimum of x and y are equal.

StaticRanges.group_max — Function

group_max(x, y[, z...])

Returns the maximum value of all collctions.

StaticRanges.group_min — Function

group_min(x, y[, z...])

Returns the minimum value of all collctions.

StaticRanges.cmpmax — Function

cmpmax(x, y)

StaticRanges.cmpmin — Function

cmpmin(x, y)

StaticRanges.min_of_group_max — Function

min_of_group_max(x, y)

Returns the minimum of maximum of x and y. Functionally equivalent to min(maximum(x), maximum(y)) but uses trait information about ordering for improved performance.

StaticRanges.max_of_group_min — Function

max_of_group_min(x, y)

Returns the maximum of minimum of x and y. Functionally equivalent to max(minimum(x), minimum(y)) but uses trait information about ordering for improved performance.

StaticRanges.is_before — Function

is_before(x::T, y::T, collection::AbstractVector{T}) -> Bool

Returns true if x is before y in collection.

is_before(x::AbstractVector{T}, y::AbstractVector{T}) -> is_before(order(x), order(y), x, y)
is_before(::Ordering, ::Ordering, x, y) -> Bool

Returns true if all elements in x are before all elements in y. Functionally equivalent to all(x .< y).

Examples

julia> using StaticRanges

julia> r1 = 1:5
1:5

julia> r2 = 6:10
6:10

julia> is_before(r2, r1)
false

julia> is_before(r1, r2)
true

StaticRanges.is_after — Function

is_after(x::T, y::T, collection::AbstractVector{T}) -> Bool

Returns true if x is after y in collection.

is_after(x::AbstractVector{T}, y::AbstractVector{T}) -> is_after(order(x), order(y), x, y)
is_after(::Ordering, ::Ordering, x, y) -> Bool

Returns true if all elements in x are after all elements in y. Functionally equivalent to all(x .> y).

Examples

julia> using StaticRanges

julia> r1 = 1:5
1:5

julia> r2 = 6:10
6:10

julia> is_after(r2, r1)
true

julia> is_after(r1, r2)
false

StaticRanges.is_contiguous — Function

is_contiguous(x, y) = is_contiguous(order(x), order(y), x, y)
is_contiguous(::Ordering, ::Ordering, x, y) -> Bool

Returns true if one of the ends of x may be extended by a single overlapping end of y.

Example

julia> using StaticRanges

julia> is_contiguous(1:3, 3:4) == true

julia> is_contiguous(3:-1:1, 3:4) == true

julia> is_contiguous(3:-1:1, 4:-1:3) == true

julia> is_contiguous(1:3, 4:-1:3) == true

julia> is_contiguous(1:3, 2:4) == false

Search and Sort

Order traits

julia> fr = 1:2:10
1:2:9

julia> rr = 10:-2:1
10:-2:2

julia> is_forward(fr)
true

julia> is_forward(rr)
false

julia> is_reverse(fr)
false

julia> is_reverse(rr)
true

These are further used to make quick simple comparisons between objects.

julia> r1 = 1:5
1:5

julia> r2 = 6:10
6:10

julia> r3 = 5:10
5:10

julia> is_before(r1, r2)  # all of r1 is before all of r2
true

julia> is_before(r1, r3)
false

julia> is_after(r2, r1)  # all of r2 is after all of r1
true

julia> is_after(r2, r3)
false

julia> is_contiguous(r1, r3)
true

julia> is_contiguous(r1, r2)
false

There are some small improvements to the family of "find and filter" methods available in base Julia. In Order to avoid type piracy but generalize the benefits to all types of ranges, some minor syntactic differences are used here.

Starting with findall, there's a difference in the type that is returned.

julia> mr = StepMRange(1, 2, 19)
StepMRange(1:2:19)

julia> @btime findall(<(5), $fr)
  142.072 ns (3 allocations: 208 bytes)
2-element Array{Int64,1}:
 1
 2

julia> @btime find_all(<(5), $fr)
  0.027 ns (0 allocations: 0 bytes)
1:2

julia> r = 1:10
1:10

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> findall(or(<(4), >(12)), mr)
7-element Array{Int64,1}:
  1
  2
  3
  7
  8
  9
 10

julia> find_all(or(<(4), >(6)), r)
7-element GapRange{Int64,UnitRange{Int64},UnitRange{Int64}}:
  1
  2
  3
  7
  8
  9
 10

julia> @btime filter(or(<(4), >(6)), $r)
  124.496 ns (3 allocations: 320 bytes)
7-element Array{Int64,1}:
  1
  2
  3
  7
  8
  9
 10

julia> @btime filter(or(<(4), >(6)), $mr)
  72.911 ns (3 allocations: 208 bytes)
7-element Array{Int64,1}:
  1
  2
  3
  7
  8
  9
 10

Experimental Syntax

Currently there's some experimental syntax that can be used to chain together functions. This functionality is an obvious case of type piracy and is currently only included as a proof of concept.

julia> findall(>(4) & <(8), fr)
2-element Array{Int64,1}:
  3
  4

julia> findall(<(4) | >(8), fr)
8-element Array{Int64,1}:
  1
  2
  5
  6
  7
  8
  9
 10

Internally it provides an intermediate structure for chaining functions an arbitrary number of functions.

julia> fxn1 = <(4) | >(8)
(::StaticRanges.ChainedFix{typeof(|),Base.Fix2{typeof(<),Int64},Base.Fix2{typeof(>),Int64}}) (generic function with 3 methods)

julia> fxn2 = <(4) | >(8) & iseven
(::StaticRanges.ChainedFix{typeof(|),Base.Fix2{typeof(<),Int64},StaticRanges.ChainedFix{typeof(&),Base.Fix2{typeof(>),Int64},typeof(iseven)}}) (generic function with 3 methods)

julia> fxn1(10)
true

julia> fxn1(11)
true

julia> fxn2(10)
true

julia> fxn2(11)
false

This becomes particularly useful when trying to preserve a range in a type stable manner. Without knowing the specific functions that compose the conditional operator in findall at compile time it's impossible to determine whether the output should be a continuous range or discrete vector.

julia> findall(i -> >(4)(i) & <(8)(i), fr)
2-element Array{Int64,1}:
 3
 4

julia> find_all(>(4) & <(8), fr)
3:4

StaticRanges.find_first — Function

find_first(predicate::Function, A)

Return the index or key of the first element of A for which predicate returns true. Return nothing if there is no such element.

Indices or keys are of the same type as those returned by keys(A) and pairs(A).

Examples

julia> using StaticRanges

julia> A = [1, 4, 2, 2];

julia> find_first(iseven, A)
2

julia> find_first(x -> x>10, A) # returns nothing, but not printed in the REPL

julia> find_first(isequal(4), A)
2

julia> find_first(iseven, [1 4; 2 2])
CartesianIndex(2, 1)

StaticRanges.find_last — Function

find_last(predicate::Function, A)

Return the index or key of the last element of A for which predicate returns true. Return nothing if there is no such element.

Indices or keys are of the same type as those returned by keys(A) and pairs(A).

Examples

julia> using StaticRanges

julia> find_last(iseven, [1, 4, 2, 2])
4

julia> find_last(x -> x>10, [1, 4, 2, 2]) # returns nothing, but not printed in the REPL

julia> find_last(isequal(4), [1, 4, 2, 2])
2

julia> find_last(iseven, [1 4; 2 2])
CartesianIndex(2, 2)

StaticRanges.merge_sort — Function

merge_sort(x, y)

Merge's and sorts collections x, and y.

StaticRanges.first_segment — Function

first_segment(x, y)

Given two collections x and y, returns the first non-overlapping segment.

StaticRanges.last_segment — Function

last_segment(x, y)

Given two collections x and y, returns the last non-overlapping segment.

StaticRanges.middle_segment — Function

middle_segment(x, y)

Given two collections x and y, returns the first non-overlapping segment.

StaticRanges.vcat_sort — Function

vcat_sort(x, y)

Returns a sorted concatenation of x and y.