Range Types

Static Ranges

StaticRanges.OneToSRange — Type

OneToSRange

A static range that parallels OneTo in behavior.

StaticRanges.UnitSRange — Type

UnitSRange

A static range parameterized by a start and stop of type T, filled with elements spaced by 1 from start until stop is exceeded. The syntax a:b with a and b both Integers creates a UnitRange.

StaticRanges.LinSRange — Type

LinSRange

A static range with len linearly spaced elements between its start and stop. The size of the spacing is controlled by len, which must be an Int.

StaticRanges.StepSRange — Type

StepSRange

A static range with elements of type T with spacing of type S. The step between each element is constant, and the range is defined in terms of a start and stop of type T and a step of type S. Neither T nor S should be floating point types.

StaticRanges.StepSRangeLen — Type

StepSRangeLen

A static range r where r[i] produces values of type T (in the second form, T is deduced automatically), parameterized by a reference value, a step, and the length. By default ref is the starting value r[1], but alternatively you can supply it as the value of r[offset] for some other index 1 <= offset <= len. In conjunction with TwicePrecision this can be used to implement ranges that are free of roundoff error.

Mutable Ranges

StaticRanges.OneToMRange — Type

OneToMRange

A mutable range that parallels OneTo in behavior.

StaticRanges.UnitMRange — Type

UnitMRange

A mutable range parameterized by a start and stop of type T, filled with elements spaced by 1 from start until stop is exceeded. The syntax a:b with a and b both Integers creates a UnitRange.

StaticRanges.LinMRange — Type

LinMRange

A mutable range with len linearly spaced elements between its start and stop. The size of the spacing is controlled by len, which must be an Int.

StaticRanges.StepMRange — Type

StepMRange

A mutable range with elements of type T with spacing of type S. The step between each element is constant, and the range is defined in terms of a start and stop of type T and a step of type S. Neither T nor S should be floating point types.

StaticRanges.StepMRangeLen — Type

StepMRangeLen

A mutable range r where r[i] produces values of type T (in the second form, T is deduced automatically), parameterized by a reference value, a step, and the length. By default ref is the starting value r[1], but alternatively you can supply it as the value of r[offset] for some other index 1 <= offset <= len. In conjunction with TwicePrecision this can be used to implement ranges that are free of roundoff error.

Abstract Ranges

StaticRanges.OneToRange — Type

OneToRange

Supertype for OneToSRange and OneToMRange. It's subtypes should behave identically to Base.OneTo.

StaticRanges.AbstractLinRange — Type

AbstractLinRange

Supertype for mutable or static ranges with len linearly spaced elements between its start and stop.

StaticRanges.AbstractStepRangeLen — Type

AbstractStepRangeLen

Supertype for StepSRangeLen and StepMRangeLen. It's subtypes should behave identically to StepRangeLen.

StaticRanges.AbstractStepRange — Type

AbstractStepRange

Supertype for StepSRange and StepMRange. It's subtypes should behave identically to StepRange.

Special Ranges

StaticRanges.GapRange — Type

GapRange{T,F,L}

Represents a range that is broken up by gaps making it noncontinuous. This allows more compact storage of numbers where the majority are known to be continuous.

Examples

julia> using StaticRanges

julia> findall(and(>(4), <(10)), 1:10)
5-element Array{Int64,1}:
 5
 6
 7
 8
 9

julia> find_all(or(<(4), >(6)), 1:10)
7-element GapRange{Int64,UnitRange{Int64},UnitRange{Int64}}:
  1
  2
  3
  7
  8
  9
 10

Manipulating Ranges

There are options for in place mutations and corresponding non mutationg operations. These allow safe mutation of ranges by avoiding states that are typically prohibited at time of construction. For example, OneToMRange cannot have a negative value for it's stop field. These methods are also called whenever setproperty! is used.

StaticRanges.can_set_first — Function

can_set_first(x) -> Bool

Returns true if the first element of x can be set. If x is a range then changing the first element will also change the length of x.

StaticRanges.set_first! — Function

set_first!(x, val)

Set the first element of x to val.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10);

julia> set_first!(mr, 2);

julia> first(mr)
2

StaticRanges.set_first — Function

set_first(x, val)

Returns similar type as x with first value set to val.

Examples

julia> using StaticRanges

julia> r = set_first(1:10, 2)
2:10

StaticRanges.can_set_step — Function

can_set_step(x) -> Bool

Returns true if type of x has step field that can be set.

StaticRanges.set_step! — Function

set_step!(x, st)

Sets the step of x to val.

Examples

julia> using StaticRanges

julia> mr = StepMRange(1, 1, 10);

julia> set_step!(mr, 2);

julia> step(mr)
2

StaticRanges.set_step — Function

set_step(x, st)

Sets the step of x to val.

Examples

julia> using StaticRanges

julia> set_step(1:1:10, 2)
1:2:9

StaticRanges.can_set_last — Function

can_set_last(x) -> Bool

Returns true if the last element of x can be set. If x is a range then changing the first element will also change the length of x.

StaticRanges.set_last! — Function

set_last!(x, val)

Set the last element of x to val.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10);

julia> set_last!(r, 5);

julia> last(mr)
5

StaticRanges.set_last — Function

set_last(x, val)

Returns a similar type as x with its last value equal to val.

Examplse

julia> using StaticRanges

julia> set_last(1:10, 5)
1:5

StaticRanges.can_set_length — Function

can_set_length(x) -> Bool

Returns true if type of x can have its length set independent of changing its first or last position.

StaticRanges.set_length! — Function

set_length!(x, len)

Returns a similar type as x with a length equal to len.

Examples

julia> using StaticRanges

julia> mr = UnitMRange(1, 10);

julia> set_length!(mr, 20);

julia> length(mr)
20

StaticRanges.set_length — Function

set_length(x, len)

Change the length of x while maintaining it's first and last positions.

Examples

julia> using StaticRanges

julia> set_length(1:10, 20)
1:20