Traits

The difference between each type of range is conceptualized as "static", "fixed", or "dynamic".

StaticRanges.is_dynamic — Function

is_dynamic(x) -> Bool

Returns true if the size of x is dynamic/can change.

Examples

julia> using StaticRanges

julia> sr = UnitSRange(1, 10)
UnitSRange(1:10)

julia> mr = StepMRange(1, 2, 20)
StepMRange(1:2:19)

julia> fr = StepRange(1, 2, 20)
1:2:19

julia> is_dynamic(sr)
false

julia> is_dynamic(fr)
false

julia> is_dynamic(mr)
true

StaticRanges.is_static — Function

is_static(x) -> Bool

Returns true if x is static.

Examples

julia> using StaticRanges

julia> sr = UnitSRange(1, 10)
UnitSRange(1:10)

julia> mr = StepMRange(1, 2, 20)
StepMRange(1:2:19)

julia> fr = StepRange(1, 2, 20)
1:2:19

julia> is_static(sr)
true

julia> is_static(mr)
false

julia> is_static(fr)
false

StaticRanges.is_fixed — Function

is_fixed(x) -> Bool

Returns true if the size of x is fixed.

Examples

julia> using StaticRanges

julia> sr = UnitSRange(1, 10)
UnitSRange(1:10)

julia> mr = StepMRange(1, 2, 20)
StepMRange(1:2:19)

julia> fr = StepRange(1, 2, 20)
1:2:19

julia> is_fixed(sr)
false

julia> is_fixed(fr)
true

julia> is_fixed(mr)
false

StaticRanges.as_dynamic — Function

as_dynamic(x)

If x is mutable then returns x, otherwise returns a comparable but mutable type to x.

Examples

julia> using StaticRanges

julia> as_dynamic(Base.OneTo(10))
OneToMRange(10)

julia> as_dynamic(UnitRange(1, 10))
UnitMRange(1:10)

julia> as_dynamic(StepRange(1, 2, 20))
StepMRange(1:2:19)

julia> as_dynamic(range(1.0, step=2.0, stop=20.0))
StepMRangeLen(1.0:2.0:19.0)

julia> as_dynamic(LinRange(1, 20, 10))
LinMRange(1.0, stop=20.0, length=10)

StaticRanges.as_static — Function

as_static(x)

If x is static then returns x, otherwise returns a comparable but static size type to x.

Examples

julia> using StaticRanges

julia> as_static(Base.OneTo(10))
OneToSRange(10)

julia> as_static(UnitRange(1, 10))
UnitSRange(1:10)

julia> as_static(StepRange(1, 2, 20))
StepSRange(1:2:19)

julia> as_static(range(1.0, step=2.0, stop=20.0))
StepSRangeLen(1.0:2.0:19.0)

julia> as_static(LinRange(1, 20, 10))
LinSRange(1.0, stop=20.0, length=10)

StaticRanges.as_fixed — Function

as_fixed(x)

If x is immutable then returns x, otherwise returns a comparable but fixed size type to x.

Examples

julia> using StaticRanges

julia> as_fixed(OneToMRange(10))
Base.OneTo(10)

julia> as_fixed(UnitMRange(1, 10))
1:10

julia> as_fixed(StepMRange(1, 2, 20))
1:2:19

julia> as_fixed(mrange(1.0, step=2.0, stop=20.0))
1.0:2.0:19.0

julia> as_fixed(LinMRange(1, 20, 10))
10-element LinRange{Float64}:
 1.0,3.11111,5.22222,7.33333,9.44444,11.5556,13.6667,15.7778,17.8889,20.0