Twice Precision

A number of operations on ranges depend on the internal type Base.TwicePrecision. This requires some unique handling for static ranges. This package has a limited number of methods that exist for consistency with base and easier readability when manipulating twice precision values internally. It's unlikely users should ever need to use these methods (or even know about them) but there's some minimal documentation included in case one should ever be so unfortunate to need to become aware of their existence.

StaticRanges.gethi — Function

gethi(x:: TwicePrecision{T}) -> T

Returns the hi component of a twice precision number.

StaticRanges.getlo — Function

getlo(x::TwicePrecision{T}) -> T

Returns the lo component of a twice precision number.

StaticRanges.stephi — Function

stephi(x::AbstractStepRangeLen)

Returns the hi component of a twice precision step.

StaticRanges.steplo — Function

steplo(x::AbstractStepRangeLen)

Returns the lo component of a twice precision step.

StaticRanges.refhi — Function

refhi(x::AbstractStepRangeLen)

Returns the hi component of a twice precision ref.

StaticRanges.reflo — Function

reflo(x::AbstractStepRangeLen)

Returns the lo component of a twice precision ref.

StaticRanges.set_ref! — Function

set_ref!(x, val)

Set the reference field of an instance of StepMRangeLen.

StaticRanges.set_offset! — Function

set_offset!(x, val)

Set the offset field of an instance of StepMRangeLen.